+++ src/core/messenger/mod.rs
@@ -125,6 +125,9 @@ impl UampApp {
             Message::Control(msg) => {
                 (Some(Message::Success), Some(Msg::Control(msg)))
             }
+            Message::Play(msg) => {
+                (Some(Message::Success), Some(Msg::PlaySong(msg)))
+            }
             _ => (Some(Message::new_error(ErrorType::ExpectedControl)), None),
         }
     }